requirements.txt
pygithub
jenkinsapi

